Separate logic structures from UI components
