+++ Apollo/Devices/Iris.cs
@@ -6,8 +6,8 @@ using System.Threading;
 
 using Newtonsoft.Json;
 
-using Apollo.Components;
 using Apollo.Elements;
+using Apollo.Structures;
 
 namespace Apollo.Devices {
     public class Iris: Device {
