+++ Apollo/Devices/Translation.cs
@@ -6,8 +6,8 @@ using System.Text;
 
 using Newtonsoft.Json;
 
-using Apollo.Components;
 using Apollo.Elements;
+using Apollo.Structures;
 
 namespace Apollo.Devices {
     public class Translation: Device {
